fix(onboard): extract inference smoke helper
